requirements.txt
streamlit==1.26.0
pandas==2.1.1
pillow==10.0.0
pyzbar==0.1.9
